refactor: Update GuestContent component and adjust imports for improved structure
- Changed import path for GuestPage to use the new GuestContent component.
- Refactored GuestContent to simplify user greeting logic and enhance button integration.
- Updated styles and structure for better readability and maintainability.
